+++ samples/gltf_viewer.cpp
@@ -63,7 +63,7 @@ using namespace filament;
 using namespace filament::math;
 using namespace filament::viewer;
 
-using namespace gltfio;
+using namespace filament::gltfio;
 using namespace utils;
 
 enum MaterialSource {